pyproject.toml
[project]
name = "paperanalyzer"
version = "0.1.0"
description = ""
authors = [
    {name = "Lorenzo Ridi"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "google-adk (>=1.19.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
